requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
python-multipart==0.0.20
httpx==0.26.0
redis==5.0.1
jinja2==3.1.3 